Handle sender and receiver connection messages
When accepting a TCP connection, we expect the first message received to
be of the form "{send|recv}|{room}". This allows the system to identify
whether the client intends to connect as a sender or a reciever, and
through which room it will be relaying data.

If the first message is not of the expected form, the client is shooed
away. We also never (currently) expect the receiver to send up anything
past the initial identification message ("recv|foo-room"), so the client
is shooed away in this case as well.