Register receiver on identification
When a client successfully identifies itself as the reciever for a room,
register it as the designated (only) receiver in the RoomRegistry. The
value in the registry is the socket identifier on which the client is
connected. This allows senders to easily look up the corresponding
receiver and send data over the socket.

Note, if multiple clients try to register as receivers for the same
room, only the first client will succeed. Later clients will be shooed
away until the first client exits.