Pipe data from sender to receiver
This accomplishes the core functionality of the relay application - data
sent from the sender side is piped to the receiver's socket.

A note on connection semantics: the reciever _must_ connect first,
before the sender attempts connecting for a specific room. If this were
not the case, a rogue receiver client may hijack a relay session by
establishing a connection for the target room before the intended client
can do so. By requiring the receiver to register first, users may be
able to detect room conflicts (either malicious or accidental) before
sending any data across the relay.