Split send and recv handlers
This patch splits the connection handler code for senders and receivers.
Previously, the code existed in a single module, but it was harder to
follow because the sender and receiver had very little in common.

By introducing a second listener dedicated to receiver clients, we can
more cleanly represent the state and lifecycle of each type of client in
its own module.