Set read timeouts
Set a read timeout of :infinity for receiver connections. This allows
them to be open as long as the sender is sending data over the pipe.
This does open the application to port starvation attacks if a malicious
(or negligent) client starts creating receivers without also completing
the other end of the pipe; this flaw should be addressed in some way.

Set a read timeout of 5 seconds for the sender - this seems reasonable.